chore: moving more ui classes
