+++ app/scripts/ui-components/CenterTiledPlot.jsx
