+++ app/scripts/ui-components/DraggableDiv.jsx
@@ -5,10 +5,10 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import clsx from 'clsx';
 
-import withTheme from './hocs/with-theme';
-import { THEME_DARK } from './configs';
+import withTheme from '../hocs/with-theme';
+import { THEME_DARK } from '../configs';
 
-import classes from '../styles/DraggableDiv.module.scss';
+import classes from '../../styles/DraggableDiv.module.scss';
 
 class DraggableDiv extends React.Component {
   constructor(props) {
