+++ app/scripts/ui-components/TiledPlot.jsx
@@ -6,26 +6,26 @@ import PropTypes from 'prop-types';
 import { ResizeSensor, ElementQueries } from 'css-element-queries';
 
 // Components
-import ContextMenuItem from './ui-components/ContextMenuItem';
+import ContextMenuItem from './ContextMenuItem';
 import CenterTrack from './CenterTrack';
 import DragListeningDiv from './DragListeningDiv';
 import GalleryTracks from './GalleryTracks';
 import TrackRenderer from './TrackRenderer';
-import { AddTrackDialog } from './ui-components';
-import CustomTrackDialog from './ui-components/CustomTrackDialog';
-import ConfigTrackMenu from './ui-components/ConfigTrackMenu';
-import CloseTrackMenu from './ui-components/CloseTrackMenu';
-import PopupMenu from './ui-components/PopupMenu';
-import ContextMenuContainer from './ui-components/ContextMenuContainer';
-import HorizontalTiledPlot from './ui-components/HorizontalTiledPlot';
+import { AddTrackDialog } from '.';
+import CustomTrackDialog from './CustomTrackDialog';
+import ConfigTrackMenu from './ConfigTrackMenu';
+import CloseTrackMenu from './CloseTrackMenu';
+import PopupMenu from './PopupMenu';
+import ContextMenuContainer from './ContextMenuContainer';
+import HorizontalTiledPlot from './HorizontalTiledPlot';
 import VerticalTiledPlot from './VerticalTiledPlot';
-import ViewContextMenu from './ui-components/ViewContextMenu';
+import ViewContextMenu from './ViewContextMenu';
 // import {HeatmapOptions} from './HeatmapOptions';
 
 // Higher-order components
-import withPubSub from './hocs/with-pub-sub';
-import withModal from './hocs/with-modal';
-import withTheme from './hocs/with-theme';
+import withPubSub from '../hocs/with-pub-sub';
+import withModal from '../hocs/with-modal';
+import withTheme from '../hocs/with-theme';
 
 // Utils
 import {
@@ -35,7 +35,7 @@ import {
   isWithin,
   sum,
   visitPositionedTracks,
-} from './utils';
+} from '../utils';
 
 // Configs
 import {
@@ -43,11 +43,11 @@ import {
   MOUSE_TOOL_SELECT,
   TRACKS_INFO_BY_TYPE,
   TRACK_LOCATIONS,
-} from './configs';
+} from '../configs';
 
 // Styles
-import styles from '../styles/TiledPlot.module.scss'; // eslint-disable-line no-unused-vars
-import stylesCenterTrack from '../styles/CenterTrack.module.scss'; // eslint-disable-line no-unused-vars
+import styles from '../../styles/TiledPlot.module.scss'; // eslint-disable-line no-unused-vars
+import stylesCenterTrack from '../../styles/CenterTrack.module.scss'; // eslint-disable-line no-unused-vars
 
 class TiledPlot extends React.Component {
   constructor(props) {
