+++ app/scripts/ui-components/TrackArea.jsx
