+++ app/scripts/ui-components/VerticalItem.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { SortableElement } from 'react-sortable-hoc';
 
-import { VerticalTrack } from './tracks';
+import { VerticalTrack } from '../tracks';
 
 const VerticalItem = SortableElement((props) => (
   <VerticalTrack
